demo/requirements.txt
django>=1.8.2,<1.9
djangorestframework>=3.1.3,<4.0.0
djangorestframework-yaml
djangorestframework-xml
djangorestframework-jsonp
